Use AnonymousCredentials directly.
The google.cloud.auth package provides a Credentials class for use with
emulators (and other systems that do not require oauth credentials). Use
this rather than rolling out own.

It seems a couple of people around the internet have copied this method
for running code against the emulators. We should ensure that if people
are going to copy these tests, that at least they do something less
hacky.