Fix tests that use _verify in the emulator.
The datastore emulator does not check the size of the response messages
before trying to send them back to the client. For all of the tests that
return large objects, this results in the response exceeding the ~4MiB
grpc message limit (enforces in message_size_filter). The response size
for these tests is ~10MiB.

Since we do not need the actual contents of the entity, we can request
just the keys, via a keys_only query.

Technically this is changing they type of query, which may be considered
testing a different thing. I will leave this up to the maintainers if
they want to merge this change.