fix: Correct indication of complete key from "partial"
Reserve_ids indicates complete_key in its typespec and in its type-checking, but says "Partial" in the description after the typespec. I also believe "use as a base for reserved IDs" is ambiguous, but I'm not familiar enough with the underlying API to know a better description since I'm not 100% sure on the behavior is supposed to be.